board - ants on grid

+++ app/src/main/java/com/tikalk/antsmasher/model/AntSpecies.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 public class AntSpecies {
 
-    private long id;
+    private int id;
     private String name;
     @ColorInt
     private int tint = Color.BLACK;
@@ -23,15 +23,15 @@ public class AntSpecies {
     public AntSpecies() {
     }
 
-    public AntSpecies(long id) {
+    public AntSpecies(int id) {
         setId(id);
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
